merged: changes from linuxport. revs: r21931-r22198
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/pvr-testing@22202 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/utils/Thread.h
@@ -57,7 +57,7 @@ public:
   CThread(IRunnable* pRunnable);
   virtual ~CThread();
   void Create(bool bAutoDelete = false, unsigned stacksize = 0);
-  DWORD ThreadId() const;
+  uintptr_t ThreadId() const;
   bool WaitForThreadExit(DWORD dwMilliseconds);
   DWORD WaitForSingleObject(HANDLE hHandle, DWORD dwMilliseconds);
   DWORD WaitForMultipleObjects(DWORD nCount, HANDLE *lpHandles, BOOL bWaitAll, DWORD dwMilliseconds);
@@ -89,7 +89,8 @@ protected:
 private:
   bool m_bAutoDelete;
   HANDLE m_StopEvent;
-  unsigned m_ThreadId;
+  unsigned m_ThreadId; // This value is unreliable on platforms using pthreads
+                       // Use m_ThreadHandle->m_hThread instead
   IRunnable* m_pRunnable;
 
   unsigned __int64 m_iLastUsage;
